+++ execUnit.hh
@@ -19,7 +19,7 @@ public:
   };
   virtual void info() = 0;
   virtual void report(std::string &s, uint64_t icnt) {}
-  virtual uint32_t getEntryAddr() const = 0;
+  virtual uint64_t getEntryAddr() const = 0;
   virtual ~execUnit() {}
   uint64_t getInscnt() const {
     return inscnt;
